backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qiskit==2.2.3
qiskit-aer==0.17.2
qiskit-ibm-runtime>=0.15.0
pydantic==2.5.0
python-dotenv==1.0.0
